Added openDragSensitivity, closeDragSensitivity, androidCloseOnBackTap, and Organized code
